backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.24.3
yfinance==0.2.18
ta==0.10.2
numpy-financial==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
vaderSentiment==3.3.2
transformers==4.35.2
plotly==5.17.0
pydantic==2.5.0
feedparser==6.0.10
Pillow==10.1.0